requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
jinja2
python-multipart
